Use one controller
