+++ src/main/java/se/iths/SpringbootGroupProject/controllers/GuestController.java
@@ -1,4 +1,4 @@
-package se.iths.SpringbootGroupProject;
+package se.iths.SpringbootGroupProject.controllers;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -10,13 +10,29 @@ import se.iths.SpringbootGroupProject.repositories.UserRepository;
 import java.util.List;
 
 @RestController
-public class MessageController {
+public class GuestController {
     final UserRepository userRepository;
     final MessageRepository messageRepository;
 
-    public MessageController(UserRepository userRepository, MessageRepository messageRepository) {
+    public GuestController(UserRepository userRepository, MessageRepository messageRepository) {
         this.userRepository = userRepository;
         this.messageRepository = messageRepository;
+    }
+
+    @GetMapping("/users")
+    List<User> findAll() {
+        return (List<User>) userRepository.findAll();
+    }
+
+    @PostMapping("/users")
+    ResponseEntity<Void> addUser(@RequestBody User user) {
+        userRepository.save(user);
+        return ResponseEntity.noContent().build();
+    }
+
+    @GetMapping("/users/{id}")
+    User findOne(@PathVariable Long id){
+        return userRepository.findById(id).get();
     }
 
     @GetMapping("/messages")
